we were being unusually strict about what selectors we would take and what we wouldn't.
we basically were saying there is no reason that *,\ should be in a selector, which just isn't true. There are hacks out there that use precisely those characters.

Since it was so arbitrary we are just going to accept those chars. We aren't very good at validating properties and I don't see much of a benefit in doing so